added 3d scene
